fix(team): correct the SCJoinedTeam wire layout and widen member ids
Ported from fd285fe4, fb8ca8cc and ff1339ae. None of these files, nor
the opcode, were touched on 10.0.2_r575:

- SCJoinedTeam is 0x10B on 10.0.2.13, not 0xd6, and its body layout was
  wrong - parties and raids never formed
- member ids are 8 bytes in SCRefreshTeamMember,
  SCTeamMemberDisconnected and SCTeamMemberRoleChanged
- party members keep their map marker once they leave visual range

The Solo* GM commands from fd285fe4 are not in this commit; they need
the ExpeditionManager and FamilyManager changes that go with the guild
work.

+++ AAEmu.Game/Core/Packets/G2C/SCTeamMemberRoleChangedPacket.cs
@@ -4,14 +4,18 @@ using AAEmu.Game.Models.Game.Team;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
-public class SCTeamMemberRoleChangedPacket(uint teamId, uint memberId, MemberRole role)
+/// <summary>
+/// Client layout (VA 0x39C773F0): tid u32, type u64, role u8. The member id is EIGHT bytes, as it is
+/// in every other team packet that carries one.
+/// </summary>
+public class SCTeamMemberRoleChangedPacket(uint teamId, ulong memberId, MemberRole role)
     : GamePacket(SCOffsets.SCTeamMemberRoleChangedPacket, 1)
 {
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(teamId);
-        stream.Write(memberId);
-        stream.Write((byte)role);
+        stream.Write(teamId);       // u32 tid
+        stream.Write(memberId);     // u64 type
+        stream.Write((byte)role);   // u8  role
         return stream;
     }
 }
